+++ common/include/NamedType.h
@@ -12,14 +12,14 @@ private:
     T value_;
 };
 
-using Temperature = NamedType<double, struct TemperatureParameter>;
-using Humidity    = NamedType<double, struct HumidityParameter   >;
+using Temperature = NamedType<float, struct TemperatureParameter>;
+using Humidity    = NamedType<float, struct HumidityParameter   >;
 
 using DHTData  = NamedType<uint8_t, struct DHTDataParameter >;
 using DHTType  = NamedType<uint8_t, struct DHTTypeParameter >;
 using DHTPower = NamedType<uint8_t, struct DHTPowerParameter>;
 
-using RadioNum   = NamedType<int8_t, struct RadioNumParameter  >;
+using RadioNum   = NamedType<int16_t, struct RadioNumParameter  >;
 using RadioPower = NamedType<int8_t, struct RadioPowerParameter>;
 using MessageId  = NamedType<uint16_t, struct MessageIdParameter>;
